requirements.txt
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langchain-chroma==0.2.2
python-dotenv==1.0.1